Add the beginnings of CloudFormer.
CloudFormer is a tool for more effectively working with CloudFormation.
The intent is to automate a lot of the more time-consuming aspects of
working with CloudFormation, such as template writing, launching stacks,
and creating AMIs from templates.

This first step introduces a cloudformer-compile-template tool. This
takes a YAML file representing the structure of a CloudFormation JSON
template. YAML is a lot cleaner and easier to work with than JSON for
this purpose (especially when dealing with multi-line data, when wanting
to add comments, etc.).

This template language also offers some additional benefits. Templates
can import other templates that contain macros for common bits of code.
They can call those macros with parameters. They can make use of a nicer
shorthand for adding references to resources/parameters and to call
CloudFormation functions. They can also define and reuse variables.

Through this, maintaining a set of CloudFormation templates will be
much easier and less error-prone.

Testing Done:
Wrote some templates utilizing all the features of the template language,
and verified the compiled templates.

Reviewed at https://rbcommons.com/s/beanbag/r/679/