Allow variables to reference other variables in the same document.
Previously, variables weren't able to reference variables created in the
same document. We only stored the variables after the document was
processed.

We now perform another variable resolution process after loading each
document, in order to finish resolving any variables that can be
resolved. This runs until we no longer have any variables left that can
be resolved. Any remaining will stay as VarReferences, for any consuming
template to deal with.

As part of this, I needed to clean up some of the state and utility
functions. There's now a TemplateState class that holds the variables,
macros, and a set of unresolved variables. It also contains the variable
resolution functions used by TemplateLoader.

This new code is also able to more intelligently handle re-constructing
a string out of a list of strings and variable references. If the
original template started as just a string instead of a list, then the
result will be a string and not a list.

Testing Done:
Unit tests pass.

Used this in some real-world templates. All variables successfully resolved
into their correct representations.

Reviewed at https://rbcommons.com/s/beanbag/r/684/