Add a command to generate Makefile-compatible dependencies files.
This adds a `cloudformer-make-depends` command that generates a
dependencies file compatible with Makefiles. This helps to make it easy
to write Makefiles that recompile CloudFormation templates whenever a
CloudFormer template or one of its dependencies have changed.

Testing Done:
Made use of this in our Makefile. Verified that the output was correct
for all templates I tested with.

Reviewed at https://rbcommons.com/s/beanbag/r/686/