Add a command and support for creating AMIs from templates.
This introduces a new command, `cloudformer-create-ami`, which will create
one or more AMIs from a template. This will compile a template, prompt
for any parameters needed by it, launch a CloudFormation stack, wait for
it to complete successfully, create AMIs from any indicated instances,
and then delete the stack.

This tool expects a new `CloudFormer` metadata section for an instance
in the template, containing a `AMINameFormat` entry. This can accept
variables and references, along with some special `{variables}` that are
handled at AMI creation time (as opposed to either compilation time or
launch time).

When compiling a template for the purpose of creating AMIs, we generate
some additional Outputs, which are values we can fetch from a completed
stack. These outputs contain all the information needed for us to
generate AMIs from the instances containing that metadata.

If the template contains a `Metadata.CloudFormer.PreviousAMI` key, and
`--update-amis-file=<filename>` is passed to `cloudformer-create-ami`, then
the given filename will have the previous AMI ID replaced with the new
ID, for each AMI created with that metadata key.

Testing Done:
Successfully created an AMI from an updated version of our ami-bootstrapper.yaml
template. Using `--update-amis-file`, I verified that it updated the old AMI
references in `amis.yaml` with the newly-generated AMI ID.

Reviewed at https://rbcommons.com/s/beanbag/r/689/