Change the style for function calls.
Functions were using `!!` prefixes, but given that there wasn't a very
good end marker for the function call, it was difficult to parse in
certain cases. It was also a bit funky when dealing with single-line
statements.

We now use `<% ... %>` for function calls. This has an added benefit of
being able to appear as a value for a key in YAML without needing to be
in quotes.

Testing Done:
Unit tests pass.

Tested with a real template.

Reviewed at https://rbcommons.com/s/beanbag/r/691/