Don't turn UncollapsibleLists into lists when processing the tree.
When processing the tree to collapse variables down, or to perform other
similar operations, we were unintentionally converting an
UncollapsibleList into a list. That caused subsequent runs to flatten
the list into a string, if that list consisted only of variables and
strings.

We now preserve the UncollapsibleList type for the resulting list in
this case.

Testing Done:
I hit this bug in a template. After this fixed, the list was preserved.

Reviewed at https://rbcommons.com/s/beanbag/r/697/