Fix processing strings with variables to references in macros.
When a macro attempted to expand variables in a string that expanded to
resource references, it would end up returning a flat list containing
all the contents, without a Fn::Join. This made CloudFormation very
grumpy.

Now, we check whether the result is a list of string items, or if it
contains anything else. If it's just strings, we combine them as before.
If it has anything else, we convert it into a Fn::Join statement.

Testing Done:
Hit this with a template I was building. This fixed it.

Unit tests pass.

Reviewed at https://rbcommons.com/s/beanbag/r/710/