Add a !cloud-init statement to help with setting cloud init data.
On EC2, the cloud-init package is used to set up the initial state of a
server. When writing a CloudFormation template, configuration data for
cloud-init can be defined in UserData. Typically, either a setup script
or a cloud-init configuration is specified. To use both, a mime message
containing both must be placed in the template, which reduces
readability and increases maintainability.

!cloud-init does this work for the template author. It accepts 'script'
and 'config' keys, which are properly processed and wrapped with all the
correct mime headers.

Testing Done:
Used this in a template to set both cloud-init configuration and to run
our bootstrapping code. Successfully deployed and verified that both had
taken effect.

Reviewed at https://rbcommons.com/s/beanbag/r/711/