Allow templates to create references based on variables.
Templates now allow for a `@@$$varname` or `@@{$$varname.path}` to
provide a reference to the name given by a variable. This is very useful
for macros, as there are plenty of constructs within CloudFormation
where you refer to a resource by name and later by reference, within the
same set of rules.

Testing Done:
Unit tests pass.

Used this in a real template and saw it expand properly.

Reviewed at https://rbcommons.com/s/beanbag/r/716/