Add a directive for embedding other files as strings.
This adds a new !embed-file directive that takes a filename, reads it, and
turns it into a joined set of lines. That makes it easy to consume files
located outside a template and store them on the server when deploying.

Testing Done:
Made use of this in a template to deploy a file.

Reviewed at https://rbcommons.com/s/beanbag/r/715/