Fix support for conditionals to better match CloudFormation's reality.
The support for conditionals previously introduced was wildly broken.
While we got valid JSON, it wasn't accepted by CloudFormation. The
reason is that If statements can't have expressions within them, but
rather must reference a defined conditional in the main "Conditions"
section of the template. This is even more of a pain for CloudFormation
template authors than I thought.

The new support for conditional expressions allows expressions to be
written naturally, and to have Condition items created automatically
from them. The resulting Fn::If will then reference that condition.

Since these conditions can be defined in a main template, an included
template, or in a macro, we can't simply populate the Conditions section
every time we find an If statement. So instead, we create placeholders
that contain the conditions, which can come along with an included
template or macro. We then post-process the tree at compile time to
generate the Conditions sections, and to replace the Fn::If statements
to reference them.

Testing Done:
Unit tests pass.

I've been deploying test stacks to AWS with real-world conditions. They've been
doing the right thing in each case I've tested.

Reviewed at https://rbcommons.com/s/beanbag/r/709/