Add a command for listing all stacks and their information.
This adds a cloudformer-list-stacks command, which lists all stacks on
CloudFormation and showing their status, creation time, description,
ARN, and outputs.

The output will be presented using colors/bold, if being written to a
terminal, to help scan the results. Passing --json will instead write
the output as pretty-printed JSON data.

Testing Done:
Tested both options, and tested outputting to a file. Saw the styling
when posting to a console, but not when outputting to a file.

Reviewed at https://rbcommons.com/s/beanbag/r/1096/