Show stack events as they come in while launching a stack.
Launching a new stack takes a while, and previously, you had to go to
the CloudFormation UI on AWS to see what was going on, or why something
had failed.

Now, those events are streamed directly to the user, in a friendly,
color-coded manner. As events come in, we show an icon (Unicode
character -- "X" for failure, checkmark for success, ">" for progress),
color-coded, and information on the event. This happens until the stack
is created or fails creation.

Testing Done:
Created several stacks, some succeeding, some failing. Compared the
events to what the CloudFormation UI was showing.

Reviewed at https://rbcommons.com/s/beanbag/r/1098/