Allow stacks to be updated with a new template or parameters.
`cloudformer-launch-stack` now accepts a `-u` argument for updating an
existing stack. When used, the `--stack-name` argument is required,
in order to specify the name of the stack to update. This also
accepts a `-k` argument, which tells it to reuse any common parameters
from the existing stack, instead of re-requesting them.

Like with creation, events from a stack update will be streamed to
the terminal. If a stack update fails, instructions will be shown
along with a command for repeating the update.

Testing Done:
Updated a couple different stacks, some succeeding, some failing. Saw
all the events and the results.

Reviewed at https://rbcommons.com/s/beanbag/r/1099/