Support creating and listing tags on stacks.
Like any other resource, a stack can have tags associated with it.
Unlike all other resources, though, these can't be set within the
template, nor can they reference parameter values. They instead have
to be created up-front when calling the API, which is pretty annoying.

We now support setting tags in the template's Meta section. These can
reference parameters, and will automatically be set when creating or
updating stacks.

The Meta section also supports taking a Name, which is the generic name
for the stack. If not provided, it's based on the filename, as before.
This value is used in the stack's GenericStackName tag.

Testing Done:
Unit tests pass.

Tested manually. Updated a stack and saw the new tags, one of which
referenced a parameter on the stack.

Reviewed at https://rbcommons.com/s/beanbag/r/1100/