Fix the create-ami command for the stack events change.
The launch-stack command was updated to print stack events, but
create-ami was never updated for the changes to the affected functions.
This updates the call to `create_stack_and_wait()` so that instead of
expecting a stack as a result, we print the events coming from it and
then look up the stack after.

Testing Done:
Re-created some AMIs and saw the stack events. AMI creation completed
successfully.

Reviewed at https://rbcommons.com/s/beanbag/r/1349/