Generate more useful error messages when failing to load templates.
If a template contains an error message, we now capture the details of
that error and display it to the user in a clear way, without simply
raising a convoluted exception.

The new `TemplateError` and `TemplateSyntaxError` subclass are used to
represent problems with templates. They're populated when loading a
template and encountering any kind fo YAML syntax error.

To further help debug issues, we also now set the name of the file in
the template reader, so that errors don't come back as being from
`<string>`.

Testing Done:
Tested this with the templates we use for our infrastructure. Saw the
syntax errors being presented in a clear way.

Unit tests passed.

Reviewed at https://rbcommons.com/s/beanbag/r/1713/