Add support for ImportValue().
This adds support for the `ImportValue()` function, allowing data from
other stacks to be referenced in a template. This is the foundation for
modern multi-stack deployments.

Testing Done:
Unit tests passed.

Used this with all the stacks we currently have in deployment.

Reviewed at https://rbcommons.com/s/beanbag/r/1721/