Fix processing of empty strings.
String processing was looking for a `__base64__` header, for converting
the whole string to Base64. This assumed there was any content in the
line, which failed if the string was empty. We now check for content
before doing this.

Testing Done:
Unit tests pass.

Tested against a template that was hitting this bug. Verified it was
fixed.

Reviewed at https://rbcommons.com/s/beanbag/r/1722/