Fix optional parameters for stack templates.
Parameters in stack templates can be optional, but our code was
continually prompting if the values weren't provided. This was an issue
also when updating existing stacks with new templates.

We now specifically check if a parameter is required, and factor that
into the prompts.

Testing Done:
Successfully created stacks with optional parameters not containing
values.

Reviewed at https://rbcommons.com/s/beanbag/r/1715/