Remove six from the codebase and fix Python 3 syntax compatibility.
In order to support modern AWS credentials, required for our deployment,
we need to move to boto3. And before we can do that, we need to move to
Python 3.

This is long-overdue. The changes required aren't terrible, but can be a
bit of a mess, so the work is being split up. To start, this change
removes all use of `six` and fixes a handful of cases where we had
syntax or operations that only worked with Python 2.

This is not a comprehensive change. Some things, like module docstrings,
have been added, but not everywhere. Further changes will iterate on the
important parts. This is not, however, going to be a full modernization
of the entire codebase. Just enough to get us back in operation.

Testing Done:
Along with other pending changes, I've successfully listed and launched
stacks and built AMIs on Python 3.

Reviewed at https://rbcommons.com/s/beanbag/r/2719/