Add typing throughout soon-to-be-updated code.
In preparation for the upcoming change to boto3, this adds typing
throughout the affected modules. This was instrumental in helping catch
some of the issues found during debugging.

This is not comprehensive. Very little of the template code, for
example, has been touched. Just enough to ensure we have proper typing
throughout the code that's going to be affected by the boto3 upgrade.

Testing Done:
Tested this along with the boto3 upgrade code.

Reviewed at https://rbcommons.com/s/beanbag/r/2721/